backend/requirements.txt
django==5.1.2
djangorestframework==3.15.2
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
reportlab==4.2.5
pillow==10.4.0
twilio==9.2.3
python-dotenv==1.0.1
gunicorn==21.2.0
django-storages[boto3]==1.14.3
